inline stuff now added
